project/screen stucture modified
