+++ lib/main.dart
@@ -1,19 +1,16 @@
 import 'package:firebase_core/firebase_core.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/svg.dart';
-import 'package:fraction/database/devilCode.dart';
-import 'package:fraction/screens/auth/register.dart';
-import 'package:fraction/screens/drawer/app_drawer.dart';
-import 'package:fraction/screens/expense/add_expense.dart';
-import 'package:fraction/screens/group/create_group.dart';
-import 'package:fraction/screens/group/group_layout.dart';
-import 'package:fraction/screens/profile/profile_layout.dart';
-import 'package:fraction/screens/auth/sign_in.dart';
-import 'package:fraction/screens/expense/view_expense_layout.dart';
+import 'package:fraction/screens/auth_layout/register.dart';
+import 'package:fraction/screens/home_layout/add_expense_view/add_expense.dart';
+import 'package:fraction/screens/create_group_layout/create_group.dart';
+import 'package:fraction/screens/home_layout/components/group_info_view.dart';
+import 'package:fraction/screens/home_layout/home_layout.dart';
+import 'package:fraction/screens/profile_layout/profile_layout.dart';
+import 'package:fraction/screens/auth_layout/sign_in.dart';
+import 'package:fraction/services/auth/auth.services.dart';
 import 'package:fraction/services/expense/expense.services.dart';
 import 'package:fraction/services/group/group.services.dart';
-import 'package:fraction/services/profile/profile.services.dart';
+import 'package:fraction/services/user/user.services.dart';
 import 'package:fraction/utils/color.dart';
 import 'package:provider/provider.dart';
 
@@ -33,7 +30,8 @@ void main() async {
     ChangeNotifierProvider(create: (context) => ApplicationState()),
     ChangeNotifierProvider(create: (context) => GroupServices()),
     ChangeNotifierProvider(create: (context) => ExpenseService()),
-    ChangeNotifierProvider(create: (context) => ProfileServices())
+    ChangeNotifierProvider(create: (context) => UserServices()),
+    ChangeNotifierProvider(create: (context) => AuthServices()),
   ], child: const MyApp()));
 }
 
@@ -82,59 +80,7 @@ class MyApp extends StatelessWidget {
                   appState.loggedIn ? const CreateGroup() : const SignInPage()),
           '/viewGroupInfo': (context) => Consumer<ApplicationState>(
               builder: (context, appState, _) =>
-                  appState.loggedIn ? const GroupLayout() : const SignInPage()),
+                  appState.loggedIn ? const GroupInfo() : const SignInPage()),
         });
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key, required this.title});
-
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  static const String clearOffIconPath = 'assets/icons/ClearOffIcon.svg';
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        title: Text(widget.title),
-        actions: [
-          kDebugMode
-              ? IconButton(
-                  onPressed: () {
-                    DevilCode()
-                        .callDevil1(shouldCallExpenseToGroupExpense: true);
-                  },
-                  icon: const Icon(Icons.refresh))
-              : Container(),
-          IconButton(
-            onPressed: () {
-              if (kDebugMode) {
-                print('');
-              }
-              // Navigator.pushNamed(context, '/profile');
-            },
-            icon: SvgPicture.asset(clearOffIconPath),
-          ),
-          const SizedBox(
-            width: 8.0,
-          )
-        ],
-      ),
-      drawer: const FractionAppDrawer(),
-      body: const ViewExpenseLayout(),
-      floatingActionButton: FloatingActionButton(
-        child: const Icon(Icons.add),
-        onPressed: () {
-          Navigator.pushNamed(context, '/addExpense');
-        },
-      ),
-    );
   }
 }
