--- lib/screens/expense/view_expense_layout.dart
@@ -1,67 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:fraction/screens/expense/components/my_dashboard.dart';
-import 'dart:math';
-
-import 'components/my_expenseview.dart';
-
-class ViewExpenseLayout extends StatefulWidget {
-  const ViewExpenseLayout({Key? key}) : super(key: key);
-
-  @override
-  createState() => ViewExpenseLayoutState();
-}
-
-class ViewExpenseLayoutState extends State<ViewExpenseLayout> {
-  final List<int> colorCodes = <int>[600, 500, 100];
-
-  String timeNow = '';
-
-  @override
-  Widget build(BuildContext context) {
-    return SingleChildScrollView(
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 16.0),
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.start,
-          mainAxisSize: MainAxisSize.max,
-          children: [
-            MyDashboard(context: context),
-            const MyExpenseView(),
-            const SizedBox(
-              height: 70,
-            )
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-final rng = Random();
-
-// all colors with shade 100
-const randomColors_shade100 = [
-  Color(0xFFBBDEFB), // blue
-  Color(0xFFC8E6C9), // green
-  Color(0xFFB2DFDB), // teal
-  Colors.red,
-  Color(0xFFFFCCBC), // deepOrange
-  Color(0xFFFFD180), // Orange
-  Colors.indigo,
-  Colors.deepPurple,
-  Colors.white,
-  Color(0xFFFFECB3), // amber
-];
-
-Color getRandomColor() {
-  return randomColors_shade200[rng.nextInt(randomColors_shade200.length)];
-}
-
-const randomColors_shade200 = [
-  //Color(0xFF80CBC4), // teal
-  Color(0xFFA5D6A7), // green
-  Color(0xFFFFE082), // amber
-  Color(0xFFCE93D8), // purple
-  Color(0xFF90CAF9), // blue
-  Color(0xFFEF9A9A), // red
-];
