+++ macos/firebase_app_id_file.json
@@ -1,7 +1,7 @@
 {
   "file_generated_by": "FlutterFire CLI",
   "purpose": "FirebaseAppID & ProjectID for this Firebase app in this directory",
-  "GOOGLE_APP_ID": "1:602053818051:ios:29cb9fe6e6e84d91f25476",
-  "FIREBASE_PROJECT_ID": "fraction-c7e1e",
-  "GCM_SENDER_ID": "602053818051"
+  "GOOGLE_APP_ID": "1:1060305160473:ios:97e21ebbf22ba5eed7dd88",
+  "FIREBASE_PROJECT_ID": "fraction-2",
+  "GCM_SENDER_ID": "1060305160473"
 }
